chore: Move useThemeMenuItems to src/hooks
