+++ superset-frontend/packages/superset-ui-core/src/components/index.ts
@@ -165,7 +165,6 @@ export * from './Table';
 export * from './TableView';
 export * from './Tag';
 export * from './TelemetryPixel';
-export * from './ThemeSubMenu';
 export * from './UnsavedChangesModal';
 export * from './constants';
 export * from './Result';
