+++ superset-frontend/src/hooks/useThemeMenuItems.tsx
